knative: Kservice: Shift helpers to util
